requirements.txt
datasets==4.0.0
pandas==2.3.1
sqlparse==0.5.3
tqdm==4.67.1
python-dotenv==1.1.1
requests==2.32.5
nltk==3.9.1
numpy==1.24.4

openai==1.35.0

# Google AI - Install at least one:
google-generativeai==0.8.5  # For Google AI Studio (API key)
# google-cloud-aiplatform>=1.38.0  # For Vertex AI (service account) - uncomment if needed

chromadb==0.4.24
sentence-transformers>=2.2.0